Crate wai_bindgen_rust

Macros

export

import

Structs

Handle

A type for handles to resources that appear in exported functions.

Traits

HandleType

A trait for types that can show up as the T in Handle<T>.

LocalHandle

An extension of the HandleType trait for locally-defined types.

Attribute Macros

async_trait